Add plots app with plotly//refactor metadata updates
