+++ app/config/settings.py
@@ -53,6 +53,8 @@ INSTALLED_APPS = [
     # Our apps
     "entities.apps.EntityConfig",
     "variables.apps.VariableConfig",
+    "plots.apps.PlotsConfig",
+
 ]
 
 MIDDLEWARE = [
@@ -71,7 +73,7 @@ ROOT_URLCONF = "config.urls"
 TEMPLATES = [
     {
         "BACKEND": "django.template.backends.django.DjangoTemplates",
-        "DIRS": [],
+        "DIRS": [BASE_DIR],
         "APP_DIRS": True,
         "OPTIONS": {
             "context_processors": [
