+++ app/plots/barchart.html
@@ -1,0 +1,11 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+  <title>Variable Dependency Chart</title>
+</head>
+<body>
+    <div>
+        <div>{{plot|safe}}</div>
+    </div>
+</body>
+</html>
