+++ app/plots/plots.py
@@ -1,0 +1,28 @@
+import numpy as np
+import plotly.graph_objects as go
+from plotly.io import to_html
+
+
+def plot1d():
+    x_data = np.arange(0, 120, 0.1)
+    trace1 = go.Scatter(
+        x=x_data,
+        y=np.sin(x_data)
+    )
+
+    data = [trace1]
+    layout = go.Layout(
+        # autosize=False,
+        # width=900,
+        # height=500,
+
+        xaxis=dict(
+            autorange=True
+        ),
+        yaxis=dict(
+            autorange=True
+        )
+    )
+    fig = go.Figure(data=data, layout=layout)
+    plot_div = fig.to_html(full_html=False)
+    return plot_div
