+++ app/plots/urls.py
@@ -1,0 +1,7 @@
+from django.urls import path
+
+from plots.views import BarChart
+
+urlpatterns = [
+    path("", BarChart)
+]
